move UserInfo type to lib and add username constructor with validation and test
